# openapi/openapi.json
{
  "swagger": "2.0",
  "paths": {
...
    "/api/presencas/eventos/{id}/gerar-qr": {
      "post": {
        "summary": "/api/presencas/eventos/{id}/gerar-qr",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
...
    "/api/presencas/eventos/{id}": {
      "get": {
        "summary": "/api/presencas/eventos/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    }
  },
...
  "info": {
    "title": "api-clube",
    "version": "1.0.0",
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3)."
  }
}